% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoar-package.R
\docType{package}
\name{spoar-package}
\alias{spoar}
\alias{spoar-package}
\title{spoar: SIMD Partial Order Alignment in R}
\description{
SPOA is a C++ library for fast alignment and consensus generation
    of closely related sequences. It allows local, global, and semi-global
    (overlap) alignment, using linear, affine, or convex gap penalties, and can
    also weight by quality scores. SPOAR is an R wrapper for SPOA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brendanf/spoar}
  \item Report bugs at \url{https://support.bioconductor.org/t/spoar}
}

}
\author{
\strong{Maintainer}: Brendan Furneaux \email{first.last@example.com} (\href{https://orcid.org/0000-0003-3522-7363}{ORCID})

Authors:
\itemize{
  \item Robert Vaser
}

}
\keyword{internal}
