% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoar.R
\name{spoaAlign}
\alias{spoaAlign}
\alias{spoaConsensus}
\title{Align sequences using SPOA (and optionally get the consensus)}
\usage{
spoaAlign(
  seq,
  m = 5L,
  n = -4L,
  g = -8L,
  e = g,
  q = g,
  c = e,
  algorithm = c("local", "global", "semi.global"),
  w = 1,
  ...
)

spoaConsensus(
  seq,
  m = 5L,
  n = -4L,
  g = -8L,
  e = g,
  q = g,
  c = e,
  algorithm = c("local", "global", "semi.global"),
  w = 1L,
  ...
)
}
\arguments{
\item{seq}{(\code{character} vector, \code{\link[Biostrings:XStringSet-class]{Biostrings::XStringSet}}, or
\code{\link[ShortRead:ShortRead-class]{ShortRead::ShortRead}}) sequences to align.}

\item{m}{(non-negative \code{integer}) score for a match. \emph{Default}: \code{5L}}

\item{n}{(non-positive \code{integer}) score for a mismatch. \emph{Default}: \code{-4L}}

\item{g}{(non-positive \code{integer}) gap opening penalty. \emph{Default}: \code{-8L}}

\item{e}{(non-positive \code{integer}) gap extension penalty. \emph{Default}: same
value as \code{g}}

\item{q}{(non-positive \code{integer}) second gap opening penalty.
\emph{Default}: same value as \code{g}}

\item{c}{(non-positive \code{integer}) second gap extension penalty.
\emph{Default}: same value as \code{e}}

\item{algorithm}{(\code{character} string) alignment mode; one of \code{"local"}
(Smith-Watterman), \code{"global"} (Needleman-Wunsch), or \code{"semi.global"}
(Overlap). \emph{Default}: \code{"local"}}

\item{w}{(\code{integer} vector) weights for the sequences. \emph{Default}: \code{1L}}

\item{...}{additional parameters passed to methods}
}
\value{
For \code{spoaConsensus()}, either a \code{character} string or the
appropriate \code{\link[Biostrings:XString-class]{Biostrings::XString}}, depending on the class of \code{seq}.

For \code{spoaAlign()}, either a
\code{character} vector or a \code{\link[Biostrings:MultipleAlignment-class]{Biostrings::MultipleAlignment}}, depending on
the class of \code{seq}. If \code{seq} is a \code{BStringSet} (i.e., an \code{XStringSet} which
is not specifically DNA, RNA, or AA) then the result is also a \code{BStringSet},
since there is no corresponding "\code{BMultipleAlignment}" class. If \code{seq} is a
\code{\link[ShortRead:ShortRead-class]{ShortRead::ShortRead}} object, the output is a
\code{\link[Biostrings:MultipleAlignment-class]{Biostrings::DNAMultipleAlignment}}.
}
\description{
Align sequences using SPOA (and optionally get the consensus)
}
\details{
\subsection{Gap penalties}{

The general (convex) gap penalty formula is:

\code{min(g + (i - i) * e, q + (i - 1) * c)}

For \code{q == g} and \code{c == e} (as is the case unless \code{q} or \code{c} is explicitly
set), this simplifies to the affine gap penalty:

\code{g + (i - 1) * e}

If, additionally, \code{e == g} (the default), then this is the linear gap
penalty:

\code{g * i}
}

\subsection{Weighting}{

Assigning a weight \emph{w} to a sequence is equivalent to including that sequence
\emph{w} times; this is primarily useful for dereplicated sequences, where all
sequences are unique, and the weight represents their multiplicity.
}

\subsection{Quality scores}{

If \code{seq} is an object which includes quality scores
(\code{\link[Biostrings:QualityScaledXStringSet-class]{Biostrings::QualityScaledXStringSet}} or \code{\link[ShortRead:ShortReadQ-class]{ShortRead::ShortReadQ}}),
then the alignment consensus is weighted using the quality scores. The method
used by SPOA uses the integer quality scores as per-position weights
analogous to sequence weights; i.e., if two different bases align in the same
position, one in a single sequence with quality score 40 and the other in a
single sequence with quality score 20, then this is equivalent to the first
base occurring at that position in 40 sequences without quality scores, and
the second base occurring at that position in 20 sequences without quality
scores.

It is possible to use a combination of sequence weights and quality scores.
In this case, in order for a dereplicated sequence set to give the same
results as the non-dereplicated sequences, the quality score for each of the
dereplicated sequences should be the mean of the quality scores for the
corresponding non-dereplicated sequences.
}
}
\examples{
sequences <- c(
    "CATAAAAGAACGTAGGTCGCCCGTCCGTAACCTGTCGGATCACCGGAAAGGACCCGTAAAGTGATAATGAT",
    "ATAAAGGCAGTCGCTCTGTAAGCTGTCGATTCACCGGAAAGATGGCGTTACCACGTAAAGTGATAATGATTAT",
    "ATCAAAGAACGTGTAGCCTGTCCGTAATCTAGCGCATTTCACACGAGACCCGCGTAATGGG",
    "CGTAAATAGGTAATGATTATCATTACATATCACAACTAGGGCCGTATTAATCATGATATCATCA",
    "GTCGCTAGAGGCATCGTGAGTCGCTTCCGTACCGCAAGGATGACGAGTCACTTAAAGTGATAAT",
    "CCGTAACCTTCATCGGATCACCGGAAAGGACCCGTAAATAGACCTGATTATCATCTACAT"
)
spoaAlign(sequences)
spoaConsensus(sequences)
}
