% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoar.R
\name{spoaAlign}
\alias{spoaAlign}
\alias{spoaConsensus}
\title{Align sequences using SPOA (and optionally get the consensus)}
\usage{
spoaAlign(
  seq,
  m = 5L,
  n = -4L,
  g = -8L,
  e = -6L,
  q = -10L,
  c = -4L,
  algorithm = c("local", "global", "semi.global"),
  gap_algorithm = c("linear", "affine", "convex"),
  ...
)

spoaConsensus(
  seq,
  m = 5L,
  n = -4L,
  g = -8L,
  e = -6L,
  q = -10L,
  c = -4L,
  algorithm = c("local", "global", "semi.global"),
  gap_algorithm = c("linear", "affine", "convex"),
  ...
)
}
\arguments{
\item{seq}{(\code{character} vector or \code{\link[Biostrings:XStringSet-class]{Biostrings::XStringSet}}) sequences
to align}

\item{m}{(non-negative \code{integer}) score for a match.
\emph{Default}: \code{5L}}

\item{n}{(non-positive \code{integer}) score for a mismatch.
\emph{Default}: \code{-4L}}

\item{g}{(non-positive \code{integer}) gap opening penalty.
\emph{Default}: \code{-8L}}

\item{e}{(non-positive \code{integer}) gap extension penalty; valid only for
\code{gap_algorithm="affine"} or \code{"convex"}.
\emph{Default}: \code{-6L}}

\item{q}{(non-positive \code{integer}) second gap opening penalty; valid only for
\code{gap_algorithm="convex"}. \emph{Default}: \code{-10L}}

\item{c}{(non-positive \code{integer}) second gap extension penalty; valid only
for \code{gap_algorithm="convex"}. \emph{Default}: \code{-4L}}

\item{algorithm}{(\code{character} string) alignment mode; one of \code{"local"}
(Smith-Watterman), \code{"global"} (Needleman-Wunsch), or \code{"semi.global"}
(Overlap). \emph{Default}: \code{"local"}}

\item{gap_algorithm}{(\code{character} string) gap scoring algorithm; one of
\code{"linear"}, \code{"affine"}, or \code{"convex"}. \emph{Default}: \code{"linear"}}

\item{...}{additional parameters passed to methods}
}
\value{
For \code{spoaConsensus()}, either a \code{character} string or the
appropriate \code{\link[Biostrings:XString-class]{Biostrings::XString}}, depending on the class of \code{seq}.

For \code{spoaAlign()}, either a
\code{character} vector or a \code{\link[Biostrings:MultipleAlignment-class]{Biostrings::MultipleAlignment}}, depending on
the class of \code{seq}. If \code{seq} is a \code{BStringSet} (i.e., an \code{XStringSet} which
is not specifically DNA, RNA, or AA) then the result is also a \code{BStringSet},
since there is no corresponding "\code{BMultipleAlignment}" class.
}
\description{
Align sequences using SPOA (and optionally get the consensus)
}
\details{
The gap penalty formulas are:

\code{linear}: \code{g * i}

\code{affine}: \code{g + (i - 1) * e}

\code{convex}: \verb{min(g + (i - i) * e, q + (i - 1) \\cdot c)}
}
\examples{
sequences <- c(
    "CATAAAAGAACGTAGGTCGCCCGTCCGTAACCTGTCGGATCACCGGAAAGGACCCGTAAAGTGATAATGAT",
    "ATAAAGGCAGTCGCTCTGTAAGCTGTCGATTCACCGGAAAGATGGCGTTACCACGTAAAGTGATAATGATTAT",
    "ATCAAAGAACGTGTAGCCTGTCCGTAATCTAGCGCATTTCACACGAGACCCGCGTAATGGG",
    "CGTAAATAGGTAATGATTATCATTACATATCACAACTAGGGCCGTATTAATCATGATATCATCA",
    "GTCGCTAGAGGCATCGTGAGTCGCTTCCGTACCGCAAGGATGACGAGTCACTTAAAGTGATAAT",
    "CCGTAACCTTCATCGGATCACCGGAAAGGACCCGTAAATAGACCTGATTATCATCTACAT"
)
spoaAlign(sequences)
spoaConsensus(sequences)
}
